+++ pkg/frontend/session.go
@@ -180,6 +180,26 @@ type Session struct {
 	planCache *planCache
 
 	autoIncrCaches defines.AutoIncrCaches
+
+	// it is for the transaction and different from the requestCtx.
+	// it is created before the transaction is started and
+	// released after the transaction is commit or rollback.
+	// the lifetime of txnCtx is longer than the requestCtx.
+	// the timeout of txnCtx is from the FrontendParameters.SessionTimeout with
+	// default 24 hours.
+	txnCtx context.Context
+}
+
+func (ses *Session) SetTxnCtx(ctx context.Context) {
+	ses.mu.Lock()
+	defer ses.mu.Unlock()
+	ses.txnCtx = ctx
+}
+
+func (ses *Session) GetTxnCtx() context.Context {
+	ses.mu.Lock()
+	defer ses.mu.Unlock()
+	return ses.txnCtx
 }
 
 // The update version. Four function.
@@ -969,273 +989,6 @@ func (ses *Session) GetConnectionID() uint32 {
 func (ses *Session) GetPeer() (string, string) {
 	rh, rp, _, _ := ses.GetMysqlProtocol().Peer()
 	return rh, rp
-}
-
-func (ses *Session) SetOptionBits(bit uint32) {
-	ses.mu.Lock()
-	defer ses.mu.Unlock()
-	ses.optionBits |= bit
-}
-
-func (ses *Session) ClearOptionBits(bit uint32) {
-	ses.mu.Lock()
-	defer ses.mu.Unlock()
-	ses.optionBits &= ^bit
-}
-
-func (ses *Session) OptionBitsIsSet(bit uint32) bool {
-	ses.mu.Lock()
-	defer ses.mu.Unlock()
-	return ses.optionBits&bit != 0
-}
-
-func (ses *Session) SetServerStatus(bit uint16) {
-	ses.mu.Lock()
-	defer ses.mu.Unlock()
-	ses.serverStatus |= bit
-}
-
-func (ses *Session) ClearServerStatus(bit uint16) {
-	ses.mu.Lock()
-	defer ses.mu.Unlock()
-	ses.serverStatus &= ^bit
-}
-
-func (ses *Session) ServerStatusIsSet(bit uint16) bool {
-	ses.mu.Lock()
-	defer ses.mu.Unlock()
-	return ses.serverStatus&bit != 0
-}
-
-/*
-InMultiStmtTransactionMode checks the session is in multi-statement transaction mode.
-OPTION_NOT_AUTOCOMMIT: After the autocommit is off, the multi-statement transaction is
-started implicitly by the first statement of the transaction.
-OPTION_BEGIN: Whenever the autocommit is on or off, the multi-statement transaction is
-started explicitly by the BEGIN statement.
-
-But it does not denote the transaction is active or not.
-
-Cases    | set Autocommit = 1/0 | BEGIN statement |
----------------------------------------------------
-Case1      1                       Yes
-Case2      1                       No
-Case3      0                       Yes
-Case4      0                       No
----------------------------------------------------
-
-If it is Case1,Case3,Cass4, Then
-
-	InMultiStmtTransactionMode returns true.
-	Also, the bit SERVER_STATUS_IN_TRANS will be set.
-
-If it is Case2, Then
-
-	InMultiStmtTransactionMode returns false
-*/
-func (ses *Session) InMultiStmtTransactionMode() bool {
-	return ses.OptionBitsIsSet(OPTION_NOT_AUTOCOMMIT | OPTION_BEGIN)
-}
-
-/*
-InActiveMultiStmtTransaction checks the session is in multi-statement transaction mode
-and there is an active transaction.
-
-But sometimes, the session does not start an active transaction even if it is in multi-
-statement transaction mode.
-
-For example: there is no active transaction.
-set autocommit = 0;
-select 1;
-
-For example: there is an active transaction.
-begin;
-select 1;
-
-When the statement starts the multi-statement transaction(select * from table), this flag
-won't be set until we access the tables.
-*/
-func (ses *Session) InActiveMultiStmtTransaction() bool {
-	return ses.ServerStatusIsSet(SERVER_STATUS_IN_TRANS)
-}
-
-/*
-TxnStart starts the transaction implicitly and idempotent
-
-When it is in multi-statement transaction mode:
-
-	Set SERVER_STATUS_IN_TRANS bit;
-	Starts a new transaction if there is none. Reuse the current transaction if there is one.
-
-When it is not in single statement transaction mode:
-
-	Starts a new transaction if there is none. Reuse the current transaction if there is one.
-*/
-func (ses *Session) TxnStart() error {
-	var err error
-	if ses.InMultiStmtTransactionMode() {
-		ses.SetServerStatus(SERVER_STATUS_IN_TRANS)
-	}
-	if !ses.GetTxnHandler().IsValidTxn() {
-		err = ses.GetTxnHandler().NewTxn()
-	}
-	return err
-}
-
-/*
-TxnCommitSingleStatement commits the single statement transaction.
-
-Cases    | set Autocommit = 1/0 | BEGIN statement |
----------------------------------------------------
-Case1      1                       Yes
-Case2      1                       No
-Case3      0                       Yes
-Case4      0                       No
----------------------------------------------------
-
-If it is Case1,Case3,Cass4, Then
-
-	InMultiStmtTransactionMode returns true.
-	Also, the bit SERVER_STATUS_IN_TRANS will be set.
-
-If it is Case2, Then
-
-	InMultiStmtTransactionMode returns false
-*/
-func (ses *Session) TxnCommitSingleStatement(stmt tree.Statement) error {
-	var err error
-	/*
-		Commit Rules:
-		1, if it is in single-statement mode:
-			it commits.
-		2, if it is in multi-statement mode:
-			if the statement is the one can be executed in the active transaction,
-				the transaction need to be committed at the end of the statement.
-	*/
-	if !ses.InMultiStmtTransactionMode() ||
-		ses.InActiveTransaction() && NeedToBeCommittedInActiveTransaction(stmt) {
-		err = ses.GetTxnHandler().CommitTxn()
-		ses.ClearServerStatus(SERVER_STATUS_IN_TRANS)
-		ses.ClearOptionBits(OPTION_BEGIN)
-	}
-	return err
-}
-
-/*
-TxnRollbackSingleStatement rollbacks the single statement transaction.
-
-Cases    | set Autocommit = 1/0 | BEGIN statement |
----------------------------------------------------
-Case1      1                       Yes
-Case2      1                       No
-Case3      0                       Yes
-Case4      0                       No
----------------------------------------------------
-
-If it is Case1,Case3,Cass4, Then
-
-	InMultiStmtTransactionMode returns true.
-	Also, the bit SERVER_STATUS_IN_TRANS will be set.
-
-If it is Case2, Then
-
-	InMultiStmtTransactionMode returns false
-*/
-func (ses *Session) TxnRollbackSingleStatement(stmt tree.Statement) error {
-	var err error
-	/*
-			Rollback Rules:
-			1, if it is in single-statement mode (Case2):
-				it rollbacks.
-			2, if it is in multi-statement mode (Case1,Case3,Case4):
-		        the transaction need to be rollback at the end of the statement.
-				(every error will abort the transaction.)
-	*/
-	if !ses.InMultiStmtTransactionMode() ||
-		ses.InActiveTransaction() {
-		err = ses.GetTxnHandler().RollbackTxn()
-		ses.ClearServerStatus(SERVER_STATUS_IN_TRANS)
-		ses.ClearOptionBits(OPTION_BEGIN)
-	}
-	return err
-}
-
-/*
-TxnBegin begins a new transaction.
-It commits the current transaction implicitly.
-*/
-func (ses *Session) TxnBegin() error {
-	var err error
-	if ses.InMultiStmtTransactionMode() {
-		ses.ClearServerStatus(SERVER_STATUS_IN_TRANS)
-		err = ses.GetTxnHandler().CommitTxn()
-	}
-	ses.ClearOptionBits(OPTION_BEGIN)
-	if err != nil {
-		/*
-			fix issue 6024.
-			When we get a w-w conflict during commit the txn,
-			we convert the error into a readable error.
-		*/
-		if moerr.IsMoErrCode(err, moerr.ErrTxnWWConflict) {
-			return moerr.NewInternalError(ses.GetRequestContext(), writeWriteConflictsErrorInfo())
-		}
-		return err
-	}
-	ses.SetOptionBits(OPTION_BEGIN)
-	ses.SetServerStatus(SERVER_STATUS_IN_TRANS)
-	err = ses.GetTxnHandler().NewTxn()
-	return err
-}
-
-// TxnCommit commits the current transaction.
-func (ses *Session) TxnCommit() error {
-	var err error
-	ses.ClearServerStatus(SERVER_STATUS_IN_TRANS | SERVER_STATUS_IN_TRANS_READONLY)
-	err = ses.GetTxnHandler().CommitTxn()
-	ses.ClearServerStatus(SERVER_STATUS_IN_TRANS)
-	ses.ClearOptionBits(OPTION_BEGIN)
-	return err
-}
-
-// TxnRollback rollbacks the current transaction.
-func (ses *Session) TxnRollback() error {
-	var err error
-	ses.ClearServerStatus(SERVER_STATUS_IN_TRANS | SERVER_STATUS_IN_TRANS_READONLY)
-	err = ses.GetTxnHandler().RollbackTxn()
-	ses.ClearOptionBits(OPTION_BEGIN)
-	return err
-}
-
-/*
-InActiveTransaction checks if it is in an active transaction.
-*/
-func (ses *Session) InActiveTransaction() bool {
-	if ses.InActiveMultiStmtTransaction() {
-		return true
-	} else {
-		return ses.GetTxnHandler().IsValidTxn()
-	}
-}
-
-/*
-SetAutocommit sets the value of the system variable 'autocommit'.
-
-The rule is that we can not execute the statement 'set parameter = value' in
-an active transaction whichever it is started by BEGIN or in 'set autocommit = 0;'.
-*/
-func (ses *Session) SetAutocommit(on bool) error {
-	if ses.InActiveTransaction() {
-		return moerr.NewInternalError(ses.requestCtx, parameterModificationInTxnErrorInfo())
-	}
-	if on {
-		ses.ClearOptionBits(OPTION_BEGIN | OPTION_NOT_AUTOCOMMIT)
-		ses.SetServerStatus(SERVER_STATUS_AUTOCOMMIT)
-	} else {
-		ses.ClearServerStatus(SERVER_STATUS_AUTOCOMMIT)
-		ses.SetOptionBits(OPTION_NOT_AUTOCOMMIT)
-	}
-	return nil
 }
 
 func (ses *Session) SetOutputCallback(callback func(interface{}, *batch.Batch) error) {
